Remove duplicate groupid and version
The group id and version of modules are not necessary if they are the
same as in the parent. Maven best practice is to omit them in a reactor
where every module has the same version.

The target platform has been left out of this change, as another PR will
remove the related pom.xml there.

Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>